Pagination query params i64-->u32

+++ deputy-package-server/src/constants.rs
@@ -1,8 +1,8 @@
-pub const fn default_page() -> i64 {
+pub const fn default_page() -> u32 {
     1
 }
 
-pub const fn default_limit() -> i64 {
+pub const fn default_limit() -> u32 {
     20
 }
